+++ src/main/java/frc/robot/AutoFactory.java
@@ -22,7 +22,7 @@ import edu.wpi.first.wpilibj2.command.WaitCommand;
 import edu.wpi.first.wpilibj2.command.sysid.SysIdRoutine;
 import frc.robot.Constants.DriveConstants;
 import frc.robot.Constants.PathConstants;
-import frc.robot.commands.driveCommands.SwerveDriveStopCommand;
+import frc.robot.commands.drivebase.SwerveDriveStopCommand;
 import frc.robot.subsystems.drive.DriveBase;
 import frc.robot.util.sysId.CharacterizableSubsystem;
