+++ src/main/java/frc/robot/commands/drivebase/SwerveDriveCommand.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands.driveCommands;
+package frc.robot.commands.drivebase;
 
 import java.util.function.BooleanSupplier;
 import java.util.function.DoubleSupplier;
