+++ src/main/java/frc/robot/commands/superstructure/ElevatorPositionCommand.java
@@ -2,19 +2,25 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands.algaeEndEffectorCommands;
+package frc.robot.commands.superstructure;
+
+import java.util.function.DoubleSupplier;
 
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.subsystems.endEffector.algae.AlgaeEndEffector;
+import frc.robot.subsystems.superstructure.Superstructure;
 
-/* You should consider using the more terse Command factories API instead https://docs.wpilib.org/en/stable/docs/software/commandbased/organizing-command-based.html#defining-commands */
-public class StopOuttakeCommand extends Command {
-  /** Creates a new StopOuttakeCommand. */
-  private final AlgaeEndEffector endEffector;
-  public StopOuttakeCommand(AlgaeEndEffector endEffector) {
+public class ElevatorPositionCommand extends Command {
+  private final Superstructure elevator;
+  private final DoubleSupplier elevatorPosition;
+  /** Creates a new ElevatorPositionCommand. */
+  public ElevatorPositionCommand(Superstructure elevator, DoubleSupplier elevatorPosition) {
+    this.elevator = elevator;
+    this.elevatorPosition = elevatorPosition;
     // Use addRequirements() here to declare subsystem dependencies.
-    this.endEffector = endEffector;
-    addRequirements(endEffector);
+    addRequirements(elevator);
+  }
+  public ElevatorPositionCommand(Superstructure elevator, double elevatorPosition) {
+    this(elevator, ()->elevatorPosition);
   }
 
   // Called when the command is initially scheduled.
@@ -24,7 +30,7 @@ public class StopOuttakeCommand extends Command {
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    endEffector.stopMotor();
+    elevator.setExtension(elevatorPosition.getAsDouble());
   }
 
   // Called once the command ends or is interrupted.
